# [MRG] Optimization for refractoriness calculation