--- a/brian2/codegen/generators/cython_generator.py
+++ b/brian2/codegen/generators/cython_generator.py
@@ -389,13 +389,7 @@ DEFAULT_FUNCTIONS['clip'].implementations.add_implementation(CythonCodeGenerator
 
 timestep_code = '''
 cdef int _timestep(double t, double dt):
-    cdef int _infinity_int = 1073741823  # maximum 32bit integer divided by 2
-    if npy_isinf(t):
-        if t < 0:
-            return -_infinity_int
-        else:
-            return  _infinity_int
-    return <int>((t + 1e-3*dt)/dt)
+    return <int64_t>((t + 1e-3*dt)/dt)
 '''
 DEFAULT_FUNCTIONS['timestep'].implementations.add_implementation(CythonCodeGenerator,
                                                                  code=timestep_code,