See discussion in #991 . The changes in this PR are:
* `timestep` function is no longer safe to use with infinity values
* `timestep` function uses `int64` to avoid overflow
* Instead of setting `lastspike` to `-inf*second` for neurons that never spiked, we now set this value to `-10000*second`
* C++ and Cython code generation has been refactored so that scalar variables are read in the scalar code block (i.e. outside of the loop), regardless of whether they are used in the scalar or vector code. This is particularly relevant for `dt` and `t`.

The last change speeds up things considerably, even though the use of `int64` slows down things again. Nevertheless, if we are looking at the time the stateupdater takes in the context of a real simulation (CUBA example), this all now looks pretty reasonable (rough "best of 3" times, do not take them too seriously):

| *target/branch*        | master+legacy        | master | This PR  |
| ------------- | ------------------:| -----------------------:| ---------------:|
| **weave**       | 126ms | 203ms | 132ms | 
| **Cython**      | 196ms | 910ms | 209ms |
| **numpy**     | 970ms | 1610ms | 1250ms |
| **C++ standalone** | 104ms | 203ms | 109ms |

"master+legacy" means activating the preference that switches off the use of the `timestep` function.