@thesamovar 2018-08-28
I think you've nailed it. Great work!